fixed shell scripts

+++ dotfilesLink.sh
@@ -2,13 +2,41 @@
 set -e
 DOT_DIRECTORY="${HOME}/dotfiles"
 
+usage() {
+  name=`basename $0`
+  cat <<EOF
+Usage:
+  $name [arguments] [command]
+Commands:
+  deploy
+Arguments:
+  -f Overwrite 
+  -h Print help
+EOF
+  exit 1
+}
+
+while getopts :fh opt; do
+  case ${opt} in
+    f)
+      OVERWRITE=true
+      ;;
+    h)
+      usage
+      ;;
+  esac
+done
+shift $((OPTIND - 1))
+
 cd ${DOT_DIRECTORY}
+
 
 link_files() {
   for f in .??*
   do
-    # Force remove the vim directory if it's already there
-    [ -n "${OVERWRITE}" -a -e ${HOME}/${f} ] && rm -f ${HOME}/${f}
+    # Force remove a dotfile if it's already there
+    [[ -f ${f} ]] &&
+      [ -n "${OVERWRITE}" -a -e ${HOME}/${f} ] && rm -f ${HOME}/${f}
     if [ ! -e ${HOME}/${f} ]; then
       # If you have ignore files, add file/directory name here
       [[ ${f} = ".git" ]] && continue
@@ -21,12 +49,12 @@ link_files() {
   echo $(tput setaf 2)Deploy dotfiles complete!. ✔︎$(tput sgr0)
 }
 
-link_files
+#link_files
 
 link_files_atom() {
   for FILE in config.cson init.coffee keymap.cson snippets.cson styles.less packages.txt
   do 
-    rm -f ${HOME}/.atom/${FILE}
+    [ -n "${OVERWRITE}" -a -e ${HOME}/.atom/${FILE} ] && rm -f ${HOME}/.atom/${FILE}
     if [ ! -e ${HOME}/.atom/${FILE} ]; then
       ln -snfv ${DOT_DIRECTORY}/.atom/${FILE} ${HOME}/.atom/${FILE}
     fi
@@ -34,15 +62,33 @@ link_files_atom() {
   echo $(tput setaf 2)Deploy atom configs complete!. ✔︎$(tput sgr0)
 }
 
-link_files_atom
+#link_files_atom
 
 # R
 link_files_R() {
   [ -n "${OVERWRITE}" -a -e ${HOME}/.R/Makevars ] && rm -f ${HOME}/.R/Makevars]
   if [ ! -e ${HOME}/.R/Makevars ]; then
-    ln -snfv ${DOT_DIRECTORY}/.R/Makevars /.R/Makevars
+    ln -snfv ${DOT_DIRECTORY}/.R/Makevars ${HOME}/.R/Makevars
   fi
   echo $(tput setaf 2)Deploy R Makevars complete!. ✔︎$(tput sgr0)
 }
 
-link_files_R
+#link_files_R
+
+command=$1
+[ $# -gt 0 ] && shift
+
+case $command in
+  deploy)
+    link_files
+    link_files_atom
+    link_files_R
+    ;;
+  #init*)
+  #  initialize
+  #  ;;
+  *)
+    usage
+    ;;
+esac
+
